stm32-viam-module/requirements.txt
googleapis-common-protos==1.59.0
grpclib==0.4.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
multidict==6.0.4
Pillow==9.4.0
protobuf==3.20.3
pyserial==3.5
typing_extensions==4.5.0
viam-sdk==0.2.14
